Move the admin helper elements to a new file
